+++ operators/ARMORED_mode_toggle.py
@@ -1,59 +1,51 @@
-# v1.0
+# v1.1
 
 import bpy
 from .. utils import addon
-# from .. utils.addon import get_prefs, get_path, get_name
 
 
 debug = False
 
-
 def main(self, context):
-    # if not bpy.ops.object.mode_set.poll or not bpy.ops.object.editmode_toggle.poll():
-    # if not bpy.ops.object.mode_set.poll():
-    #     print('ARMTK - Operator Poll Failed')
-    #     self.report({'DEBUG'}, 'ARMORED Toolkit\n Sub-Operator Poll Failed')
-    #     return {'FINISHED'}
+    # if bpy.ops.object.mode_set.poll():
+    if bpy.ops.object.editmode_toggle.poll():
+        bpy.ops.object.mode_set(mode='EDIT', toggle=True)
 
-    workspace = (context.window.workspace.name_full).upper()
-    # bpy.ops.object.mode_set(toggle=True)
+    # if bpy.ops.object.editmode_toggle.poll():
+        # bpy.ops.object.editmode_toggle()
 
+# def main(self, context):
+#     workspace = (context.window.workspace.name_full).upper()
 
-
-    if context.object.mode == 'SCULPT':
-        # bpy.ops.object.mode_set(mode='OBJECT', toggle=True)
-        bpy.ops.sculpt.sculptmode_toggle()  # Will toggle to Object Mode.
-        if debug: print('Sculp mode detected, toggle to object? \n')
+#     if context.object.mode == 'SCULPT':
+#         bpy.ops.sculpt.sculptmode_toggle()  # Will toggle to Object Mode.
+#         if debug: print('Sculp mode detected, toggle to object? \n')
     
-    elif context.object.type == 'GPENCIL':
-        # bpy.ops.object.mode_set(mode='EDIT_GPENCIL', toggle=True)
-        bpy.ops.gpencil.editmode_toggle()   # Will toggle between Object end Edit GPencil.
-        if debug: print('GPencil object type detected, toggle between EDIT/OBJECT \n')
+#     elif context.object.type == 'GPENCIL':
+#         bpy.ops.gpencil.editmode_toggle()   # Will toggle between Object end Edit GPencil.
+#         if debug: print('GPencil object type detected, toggle between EDIT/OBJECT \n')
     
-    elif workspace == 'SCULPTING':
-        if debug: print('Sculpt Worspace detected')
-        try:
-            # bpy.ops.object.mode_set(mode='SCULPT', toggle=True)
-            bpy.ops.sculpt.sculptmode_toggle()      # Will toggle between Object end Sculpt.
-            if debug: print('Toggle to sculpt mode successful \n')
+#     elif workspace == 'SCULPTING':
+#         if debug: print('Sculpt Worspace detected')
+#         try:
+#             bpy.ops.sculpt.sculptmode_toggle()      # Will toggle between Object end Sculpt.
+#             if debug: print('Toggle to sculpt mode successful \n')
 
-        except TypeError:
-        # except RuntimeError:
-            if debug: print('Runtime Error')
-            if bpy.ops.object.editmode_toggle.poll():
-                bpy.ops.object.editmode_toggle()        # Will toggle between Object end Edit.
-                # bpy.ops.object.mode_set(mode='EDIT', toggle=True)
-                if debug: print('Toggle to edit mode successful \n')
-            else:
-                if debug: print('Edit mode Poll Failed \n')
+#         except TypeError:
+#         # except RuntimeError:
+#             if debug: print('Runtime Error')
+#             if bpy.ops.object.editmode_toggle.poll():
+#                 bpy.ops.object.editmode_toggle()        # Will toggle between Object end Edit.
+#                 if debug: print('Toggle to edit mode successful \n')
+#             else:
+#                 if debug: print('Edit mode Poll Failed \n')
 
-    else:
-        if bpy.ops.object.editmode_toggle.poll():
-            # bpy.ops.object.mode_set(mode='EDIT', toggle=True)
-            bpy.ops.object.editmode_toggle()        # Will toggle between Object end Edit.
-            if debug: print('Toggle to edit mode successful \n')
-        else:
-            if debug: print('Edit mode Poll Failed \n')
+#     else:
+#         if bpy.ops.object.editmode_toggle.poll():
+#             bpy.ops.object.editmode_toggle()        # Will toggle between Object end Edit.
+#             if debug: print('Toggle to edit mode successful \n')
+#         else:
+#             if debug: print('Edit mode Poll Failed \n')
 
 
 
@@ -100,17 +92,7 @@ class ARMORED_OT_mode_toggle_none(bpy.types.Operator):
 
 
 
-# classes = (
-    # ARMORED_OT_mode_toggle,
-    # ARMORED_OT_mode_toggle_wrapper,
-    # ARMORED_OT_mode_toggle_standard,
-    # ARMORED_OT_mode_toggle_grouped_undo,
-    # ARMORED_OT_mode_toggle_no_undo,
-# )
-
 def register():
-    # from .. utils.preferences import get_prefs, get_path, get_name
-
     undo_mode = addon.preferences().tab_undo_mode
     # print(f'TAB Undo History Mode: {undo_mode}')
 
@@ -122,6 +104,7 @@ def register():
 
     elif undo_mode == 'NONE':
         bpy.utils.register_class(ARMORED_OT_mode_toggle_none)
+
 
 def unregister():